"Add description field to income form and update display logic"
This commit adds a new 'description' field to the 'income' form. It updates the back-end logic in 'views.py' to accept, save, and return this new field. The front-end code has been updated to handle the description input in the form and display a character counter. In addition, entries with a description are now marked in the calendar view with a special indicator. This will help users to better understand their income entries by providing additional context.