"Implement currency formatting for income display"
In this commit, we've introduced a new currency formatting function for the income display. This function transforms amounts greater than or equal to 1000 from the full numerical representation to a simplified format using a 'k' suffix to denote thousands. For instance, an amount of 1500 PLN is now displayed as 1.5k PLN. This change is implemented in both the Python code and in the JavaScript used on the front-end.

The goal of this change is to improve readability and usability by simplifying larger numbers.