"Improve expense calendar UI and add detailed daily expenses JSON"
This commit enhances the expense calendar UI by applying new styling for a better user experience. It also includes a significant change to the daily_expense_list in the expense_calendar view. Previously, this was a simple list of totals per day, but it is now a dictionary of JSON strings representing detailed expenses per day. This change will allow the frontend to display detailed daily expense information, providing more insight for the user.